Added two main views and set up MVVM architecture and navigation
