+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Mon Dec 13 13:26:12 MST 2021
+gradle.version=7.0.2
