+++ app/src/main/java/com/capwaydemo/NewsFeedRepository.kt
@@ -1,0 +1,4 @@
+package com.capwaydemo
+
+class NewsFeedRepository {
+}
