+++ app/src/main/java/com/capwaydemo/Utils.kt
@@ -1,0 +1,5 @@
+package com.capwaydemo
+
+object Utils {
+    const val KEY_EMAIL_OR_PHONE = "KEY_EMAIL_OR_PHONE"
+}
