+++ app/src/main/java/com/capwaydemo/ui/MainActivity.kt
@@ -1,0 +1,12 @@
+package com.capwaydemo.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.capwaydemo.R
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
