+++ app/src/main/java/com/capwaydemo/ui/NewsFeed/NewsFeedViewModel.kt
@@ -1,0 +1,12 @@
+package com.capwaydemo.ui.NewsFeed
+
+import androidx.lifecycle.ViewModel
+
+class NewsFeedViewModel() : ViewModel() {
+
+    fun getNewsFeed(searchCriteria: String) {
+
+    }
+
+
+}
