--- app/src/main/java/com/capwaydemo/ui/home/HomeViewModel.kt
@@ -1,13 +1,0 @@
-package com.capwaydemo.ui.home
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
-    }
-    val text: LiveData<String> = _text
-}
