+++ app/src/main/res/navigation/main_navigation.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_navigation"
+    app:startDestination="@+id/nav_news_feed">
+
+    <fragment
+        android:id="@+id/nav_news_feed"
+        android:name="com.capwaydemo.ui.NewsFeed.NewsFeedFragment"
+        android:label="@string/menu_news_feed"
+        tools:layout="@layout/fragment_news_feed">
+
+        <action
+            android:id="@+id/action_news_feed_to_details"
+            app:destination="@id/nav_article_details"
+            app:enterAnim="@anim/nav_default_enter_anim"
+            app:exitAnim="@anim/nav_default_exit_anim"
+            app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+            app:popExitAnim="@anim/nav_default_pop_exit_anim" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/nav_article_details"
+        android:name="com.capwaydemo.ui.FeedDetails.ArticleDetailsFragment"
+        android:label="@string/menu_article_details"
+        tools:layout="@layout/fragment_article_details" />
+</navigation>
