+++ app/src/main/res/values/strings.xml
@@ -7,7 +7,6 @@
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
 
-    <string name="menu_home">Home</string>
-    <string name="menu_gallery">Gallery</string>
-    <string name="menu_slideshow">Slideshow</string>
+    <string name="menu_news_feed">News Feed</string>
+    <string name="menu_article_details">Article Details</string>
 </resources>
