+++ app/src/main/res/values/themes.xml
@@ -1,25 +1,88 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.CapWayDemo" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
-        <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/purple_500</item>
-        <item name="colorPrimaryVariant">@color/purple_700</item>
+
+    <style name="MainTheme" parent="@style/Theme.MaterialComponents.Light.NoActionBar">
+        <item name="android:windowNoTitle">true</item>
+        <item name="android:windowActionBar">false</item>
+        <item name="android:windowContentOverlay">@null</item>
+
+        <item name="colorPrimary">@color/primary</item>
+        <item name="android:colorPrimaryDark">@color/black</item>
         <item name="colorOnPrimary">@color/white</item>
+
         <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/teal_200</item>
-        <item name="colorSecondaryVariant">@color/teal_700</item>
+        <item name="colorPrimaryVariant">@color/primary</item>
+        <item name="colorSecondary">@color/primary</item>
+        <item name="colorSecondaryVariant">@color/primary</item>
         <item name="colorOnSecondary">@color/black</item>
-        <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
-        <!-- Customize your theme here. -->
     </style>
 
-    <style name="Theme.CapWayDemo.NoActionBar">
-        <item name="windowActionBar">false</item>
-        <item name="windowNoTitle">true</item>
-    </style>
 
-    <style name="Theme.CapWayDemo.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
 
-    <style name="Theme.CapWayDemo.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
+<!--    <style name="primaryButton">-->
+
+<!--        <item name="android:fontFamily">@font/graphik_medium</item>-->
+<!--        <item name="android:textAllCaps">false</item>-->
+<!--        <item name="android:textSize">16sp</item>-->
+<!--        <item name="android:gravity">center</item>-->
+<!--        <item name="android:letterSpacing">0</item>-->
+<!--        <item name="android:backgroundTint">@color/primary</item>-->
+<!--        <item name="android:textColor">@color/white</item>-->
+<!--    </style>-->
+
+<!--    <style name="primaryTextButton" parent="Widget.MaterialComponents.Button.TextButton" >-->
+<!--        <item name="android:fontFamily">@font/graphik_medium</item>-->
+<!--        <item name="android:buttonStyle">@style/Widget.MaterialComponents.Button.TextButton</item>-->
+<!--        <item name="android:textAllCaps">false</item>-->
+<!--        <item name="android:textSize">16sp</item>-->
+<!--        <item name="android:gravity">center</item>-->
+<!--        <item name="android:letterSpacing">0</item>-->
+<!--        <item name="android:backgroundTint">@color/primary</item>-->
+<!--        <item name="android:textColor">@color/white</item>-->
+<!--        <item name="android:paddingTop">12dp</item>-->
+<!--        <item name="android:paddingBottom">12dp</item>-->
+<!--    </style>-->
+
+
+<!--    <style name="titleLarge">-->
+<!--        <item name="android:fontFamily">@font/graphik_regular</item>-->
+<!--        <item name="android:textAllCaps">false</item>-->
+<!--        <item name="android:textSize">35sp</item>-->
+<!--        <item name="android:gravity">center</item>-->
+<!--        <item name="android:textColor">@color/black</item>-->
+<!--    </style>-->
+
+<!--    <style name="titleMedium">-->
+<!--        <item name="android:fontFamily">@font/graphik_regular</item>-->
+<!--        <item name="android:textAllCaps">false</item>-->
+<!--        <item name="android:textSize">24sp</item>-->
+<!--        <item name="android:gravity">center</item>-->
+<!--        <item name="android:textColor">@color/black</item>-->
+<!--    </style>-->
+
+<!--    <style name="body">-->
+<!--        <item name="android:fontFamily">@font/graphik_regular</item>-->
+<!--        <item name="android:textAllCaps">false</item>-->
+<!--        <item name="android:textSize">14sp</item>-->
+<!--        <item name="android:lineSpacingMultiplier">1.2</item>-->
+<!--        <item name="android:gravity">center</item>-->
+<!--        <item name="android:textColor">@color/black</item>-->
+<!--    </style>-->
+
+<!--    <style name="headerLarge">-->
+<!--        <item name="android:fontFamily">@font/graphik_semibold</item>-->
+<!--        <item name="android:textAllCaps">false</item>-->
+<!--        <item name="android:textSize">24sp</item>-->
+<!--        <item name="android:gravity">center</item>-->
+<!--        <item name="android:textColor">@color/black</item>-->
+<!--    </style>-->
+
+<!--    <style name="label">-->
+<!--        <item name="android:fontFamily">@font/graphik_semibold</item>-->
+<!--        <item name="android:textAllCaps">true</item>-->
+<!--        <item name="android:textSize">16sp</item>-->
+<!--        <item name="android:gravity">center</item>-->
+<!--        <item name="android:textColor">@color/black</item>-->
+<!--    </style>-->
+
 </resources>
